pyproject.toml
[tool.poetry]
name = "ApplicationDistributionCenter"
version = "0.1.0"
description = ""
authors = ["VincentAdamNemessis <vincentadamnemessis@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[virtualenvs]
create=true
in-project=true

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
django-simpleui = "^2024.10.25"
django-redis = "^5.4.0"
scikit-learn = "^1.5.2"
markdown = "^3.7"
cryptography = "^43.0.3"
pycryptodome = "^3.21.0"
comtypes = "^1.4.8"
pillow = "^11.0.0"
redis = "^5.2.0"
jieba = "^0.42.1"
pyecharts = "^2.0.7"
django-import-export = "^4.2.0"
bs4 = "^0.0.2"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
loguru = "^0.7.2"
more-itertools = "^10.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
priority = "supplemental"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "supplemental"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "alembic",
]


line-length = 120
indent-width = 4

target-version = "py310"

[tool.ruff.lint]
select = [
    "A",  # flake8-annotations
    "B", # flake8-bugbear rules
    "F", # pyflakes rules
    "N",  #  name style rules
    "I", # isort rules
    "UP",   # pyupgrade rules
    "E101", # mixed-spaces-and-tabs
    "E111", # indentation-with-invalid-multiple
    "E112", # no-indented-block
    "E113", # unexpected-indentation
    "E115", # no-indented-block-comment
    "E116", # unexpected-indentation-comment
    "E117", # over-indented
    "RUF019", # unnecessary-key-check
    "RUF100", # unused-noqa
    "RUF101", # redirected-noqa
    "S506", # unsafe-yaml-load
    "W191", # tab-indentation
    "W605", # invalid-escape-sequence
]
ignore = [
    "B904", # raise-without-from-inside-except
    "N999" # ignore name style rule
]

fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_[a-zA-Z0-9_]*|)$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false


[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = []

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.pylint]

disable = ["all"]  # diable all rule first
enable = ["too-many-statements"]  # then enable too-many-statements rule
max-statements = 50  # function max statement